Corrections on the comments from W. E. Brown.
Aside from typos and mistakes, prohibited non-stable-layout members and base
classes in a stable-layout class. Mostly this is due the fact that the standard
does not require types that are not trivially-copyable and not standard-layout
to have contiguous storage, and I found it difficult to produce a motivating
example to extend that guarantee to classes with virtual functions and virtual
base classes. The main motivating example of the proposal also does not require
such types, so it appeared easier to drop them. In return, the proposed wording
got much simpler.

Also added a type trait to test if a type is a stable-layout class.